Rename implementation types for consistency
Implements #24
