+++ src/com/amazon/ionelement/impl/BlobElementImpl.kt
@@ -16,23 +16,24 @@
 package com.amazon.ionelement.impl
 
 import com.amazon.ion.IonWriter
-import com.amazon.ionelement.api.ClobElement
-import com.amazon.ionelement.api.ElementType
+import com.amazon.ionelement.api.BlobElement
 import com.amazon.ionelement.api.ByteArrayView
+import com.amazon.ionelement.api.ElementType
 import com.amazon.ionelement.api.MetaContainer
 import com.amazon.ionelement.api.emptyMetaContainer
 
-internal class ClobIonElement(
+internal class BlobElementImpl(
     bytes: ByteArray,
     override val annotations: List<String> = emptyList(),
     override val metas: MetaContainer = emptyMetaContainer()
-) : BinaryIonElement(bytes), ClobElement {
+) : LobElementBase(bytes), BlobElement {
 
-    override val clobValue: ByteArrayView get() = bytesValue
+    override val blobValue: ByteArrayView get() = bytesValue
 
-    override fun writeContentTo(writer: IonWriter) = writer.writeClob(bytes)
-    override fun copy(annotations: List<String>, metas: MetaContainer): ClobElement =
-        ClobIonElement(bytes, annotations, metas)
+    override fun writeContentTo(writer: IonWriter) = writer.writeBlob(bytes)
 
-    override val type: ElementType get() = ElementType.CLOB
+    override fun copy(annotations: List<String>, metas: MetaContainer): BlobElement =
+        BlobElementImpl(bytes, annotations, metas)
+
+    override val type: ElementType get() = ElementType.BLOB
 }
