+++ src/com/amazon/ionelement/impl/LongIntElementImpl.kt
@@ -16,32 +16,31 @@
 package com.amazon.ionelement.impl
 
 import com.amazon.ion.IonWriter
-import com.amazon.ion.Timestamp
 import com.amazon.ionelement.api.ElementType
+import com.amazon.ionelement.api.IntElement
+import com.amazon.ionelement.api.IntElementSize
 import com.amazon.ionelement.api.MetaContainer
-import com.amazon.ionelement.api.TimestampElement
 import com.amazon.ionelement.api.emptyMetaContainer
 
-internal class TimestampIonElement(
-    val value: Timestamp,
+internal class LongIntElementImpl(
+    override val longValue: Long,
     override val annotations: List<String> = emptyList(),
     override val metas: MetaContainer = emptyMetaContainer()
-): AnyElementBase(), TimestampElement {
-    constructor(timestamp: String): this(Timestamp.valueOf(timestamp))
+) : AnyElementBase(), IntElement {
+    override val integerSize: IntElementSize get() = IntElementSize.LONG
+    override val type: ElementType get() = ElementType.INT
 
-    override val type: ElementType get() = ElementType.TIMESTAMP
-    override fun copy(annotations: List<String>, metas: MetaContainer): TimestampElement =
-        TimestampIonElement(value, annotations, metas)
+    override fun copy(annotations: List<String>, metas: MetaContainer): IntElement =
+        LongIntElementImpl(longValue, annotations, metas)
 
-    override fun writeContentTo(writer: IonWriter) = writer.writeTimestamp(value)
-
+    override fun writeContentTo(writer: IonWriter) = writer.writeInt(longValue)
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
 
-        other as TimestampIonElement
+        other as LongIntElementImpl
 
-        if (value != other.value) return false
+        if (longValue != other.longValue) return false
         if (annotations != other.annotations) return false
         // Note: metas intentionally omitted!
 
@@ -49,9 +48,10 @@ internal class TimestampIonElement(
     }
 
     override fun hashCode(): Int {
-        var result = value.hashCode()
+        var result = longValue.hashCode()
         result = 31 * result + annotations.hashCode()
         // Note: metas intentionally omitted!
         return result
     }
 }
+
