+++ src/com/amazon/ionelement/impl/TextElementBase.kt
@@ -18,7 +18,7 @@ package com.amazon.ionelement.impl
 import com.amazon.ionelement.api.MetaContainer
 import com.amazon.ionelement.api.emptyMetaContainer
 
-internal abstract class TextIonElement(
+internal abstract class TextElementBase(
     override val textValue: String,
     override val annotations: List<String> = emptyList(),
     override val metas: MetaContainer = emptyMetaContainer()
